improved 1x14 header, M48T59Y-MH footprints
The 1x14 header footprint is for round (machine pin)
headers, which is what we want to use. The through holes
are much more compact than the footprint for normal
square headers.

The M48T59Y footprint is made slightly more compact.
It should still allow plenty of room to solder on the
M49T59Y, but it fits relatively comfortably between
the 1x14 pin headers.

With these changes, it shouldn't be necessary to bend
the legs of the M48T59Y.