implement reads and writes
It turns out that not all of the pins of PORTC are
available on the Arduino Nano, so changed the test
circuit and test program to use the low 3 bits of
PORTB to interface with the high nybble of the
data bus.

I can successfully write to and read from two bytes
of memory on a DIP M48T59Y, which is good progress.