major refactoring
Was trying to figure out why all tests will failing
with every M48T59Y, but the problem actually seems
to be that I switched the Arduino to a USB hub which
was not producing a sufficiently high VCC! Plugging
the Arduino directly into one of the computer's front
panel USB ports seems to have resolved the issue.

Grr, lost half a day to this nonsense.